[Agent] add shared entityRef resolution util
