+++ src/logic/operationHandlers/addComponentHandler.js
@@ -13,6 +13,7 @@
 /** @typedef {import('./modifyComponentHandler.js').EntityRefObject} EntityRefObject */ // Reuse definition
 /** @typedef {import('../../interfaces/ISafeEventDispatcher.js').ISafeEventDispatcher} ISafeEventDispatcher */
 import { SYSTEM_ERROR_OCCURRED_ID } from '../../constants/eventIds.js';
+import resolveEntityId from '../../utils/entityRefUtils.js';
 
 /**
  * Parameters accepted by {@link AddComponentHandler#execute}.
@@ -69,35 +70,6 @@ class AddComponentHandler {
   }
 
   /**
-   * Resolves entity_ref -> entityId or null.
-   * (Copied directly from ModifyComponentHandler as the logic is identical)
-   *
-   * @private
-   * @param {AddComponentOperationParams['entity_ref']} ref - The entity reference from parameters.
-   * @param {ExecutionContext} ctx - The execution context.
-   * @returns {string | null} The resolved entity ID or null.
-   */
-  #resolveEntityId(ref, ctx) {
-    const ec = ctx?.evaluationContext ?? {};
-    if (typeof ref === 'string') {
-      const t = ref.trim();
-      if (!t) return null;
-      if (t === 'actor') return ec.actor?.id ?? null;
-      if (t === 'target') return ec.target?.id ?? null;
-      return t; // Assume direct ID
-    }
-    if (
-      ref &&
-      typeof ref === 'object' &&
-      typeof ref.entityId === 'string' &&
-      ref.entityId.trim()
-    ) {
-      return ref.entityId.trim();
-    }
-    return null;
-  }
-
-  /**
    * Executes the ADD_COMPONENT operation.
    * Adds a new component instance (or replaces an existing one) on the specified entity.
    *
@@ -138,7 +110,7 @@ class AddComponentHandler {
     const trimmedComponentType = component_type.trim();
 
     // 2. Resolve Entity ID
-    const entityId = this.#resolveEntityId(entity_ref, executionContext);
+    const entityId = resolveEntityId(entity_ref, executionContext);
     if (!entityId) {
       log.warn(`ADD_COMPONENT: Could not resolve entity id from entity_ref.`, {
         entity_ref,
